Use intmax_t type and PRIdMAX to print format pid_t pids
We don't know how large pid_t can be, so cast it to intmax_t and
use PRIdMAX to ensure we have large enought int formatting to cover
the full pid_t range.

+++ core-ftrace.c
@@ -280,7 +280,7 @@ void stress_ftrace_add_pid(const pid_t pid)
 	if (pid == -1) {
 		strcpy(buffer, " ");
 	} else {
-		snprintf(buffer, sizeof(buffer), "%d", (int)pid);
+		snprintf(buffer, sizeof(buffer), "%" PRIdMAX , (intmax_t)pid);
 	}
 	ret = write(fd, buffer, strlen(buffer));
 	(void)ret;
